Select link updates

+++ support/reporter.ts
@@ -24,15 +24,17 @@ export class Reporter {
         public static cleanseDirectory(dir: string) {
             const fs = require('fs');
             const path = require('path');
-            fs.readdir(dir, (err, files) => {
-                if (err) throw err;
+            if (fs.existsSync(dir)) {
+                fs.readdir(dir, (err, files) => {
+                    if (err) throw err;
 
-            for (const file of files) {
-            fs.unlink(path.join(dir, file), err => {
-                if (err) throw err;
+                for (const file of files) {
+                fs.unlink(path.join(dir, file), err => {
+                    if (err) throw err;
                 });
             }
             });
+        }
             }
